✨ feat(backend): Add auth user and test

+++ backend/api/v0/routes/users/controller.py
@@ -10,11 +10,15 @@ It includes endpoints for:
 
 from fastapi import APIRouter
 from fastapi import Depends
+from fastapi import HTTPException
 from fastapi import Response
+from fastapi import status
 
 from api.model import FieldSelectionParams
 from api.model import PaginatedResponse
 from utils.dependency.initialization import db_dep
+from utils.dependency.initialization import user_admin_dep
+from utils.dependency.initialization import user_dep
 
 from . import model
 from . import service
@@ -30,6 +34,7 @@ router = APIRouter(prefix="/v0/users", tags=["Users"])
 )
 def get_all_users(
     db: db_dep,
+    user_token: user_dep,
     response: Response,
     params: model.AllUsersQueryParams = Depends(),
 ):
@@ -37,6 +42,7 @@ def get_all_users(
 
     Args:
         db (db_dep): The database session.
+        user_token (user_dep): The user session.
         response (Response): The FastAPI response object to set status codes.
         params (model.AllUsersQueryParams): Query parameters for filtering
             users.
@@ -45,6 +51,13 @@ def get_all_users(
         PaginatedResponse[model.UserResponse]: A list of user data and
             pagination metadata.
     """
+    if not user_token:
+        raise HTTPException(
+            status_code=status.HTTP_401_UNAUTHORIZED,
+            detail="Could not validate credentials",
+            headers={"WWW-Authenticate": "Bearer"},
+        )
+
     all_users = service.get_all_users(db, response)
 
     # Filter
@@ -101,17 +114,29 @@ def get_all_users(
 
 
 @router.post("/", response_model=model.UserResponse, status_code=201)
-def create_user(user: model.UserCreate, db: db_dep, response: Response):
+def create_user(
+    user: model.UserCreate,
+    db: db_dep,
+    user_token: user_admin_dep,
+    response: Response,
+):
     """Endpoint to create a new user.
 
     Args:
         user (model.UserCreate): The user data to create.
         db (db_dep): The database session.
+        user_token (user_admin_dep): The user session with admin scope.
         response (Response): The FastAPI response object to set status codes.
 
     Returns:
         model.UserResponse: The created user data.
     """
+    if user_token.scopes != "admin":
+        raise HTTPException(
+            status_code=status.HTTP_401_UNAUTHORIZED,
+            detail="Not enough permissions",
+            headers={"WWW-Authenticate": "Bearer scope=admin"},
+        )
     return service.create_user(user, db, response)
 
 
@@ -124,6 +149,7 @@ def create_user(user: model.UserCreate, db: db_dep, response: Response):
 def get_user(
     id: str,
     db: db_dep,
+    user_token: user_dep,
     response: Response,
     params: FieldSelectionParams = Depends(),
 ):
@@ -139,6 +165,13 @@ def get_user(
     Returns:
         model.UserResponse: The user data if found.
     """
+    if not user_token:
+        raise HTTPException(
+            status_code=status.HTTP_401_UNAUTHORIZED,
+            detail="Could not validate credentials",
+            headers={"WWW-Authenticate": "Bearer"},
+        )
+
     user = service.get_user_byid(id, db, response)
     # Field Selection
     if params.fields:
@@ -155,7 +188,11 @@ def get_user(
 
 @router.put("/{id}", response_model=model.UserResponse)
 def update_user(
-    id: str, user: model.UserUpdate, db: db_dep, response: Response
+    id: str,
+    user: model.UserUpdate,
+    db: db_dep,
+    user_token: user_dep,
+    response: Response,
 ):
     """Endpoint to update an existing user's information.
 
@@ -163,40 +200,69 @@ def update_user(
         id (str): The unique identifier of the user to update.
         user (model.UserUpdate): The updated user data.
         db (db_dep): The database session.
+        user_token (user_admin_dep): The user session with admin scope.
         response (Response): The FastAPI response object to set status codes.
 
     Returns:
         model.UserResponse: The updated user data.
     """
+    if user_token.scopes != "admin":
+        raise HTTPException(
+            status_code=status.HTTP_401_UNAUTHORIZED,
+            detail="Not enough permissions",
+            headers={"WWW-Authenticate": "Bearer scope=admin"},
+        )
     return service.update_user(id, user, db, response)
 
 
 @router.patch("/{id}", response_model=model.UserResponse, status_code=200)
-def patch_user(id: str, user: model.UserPatch, db: db_dep, response: Response):
+def patch_user(
+    id: str,
+    user: model.UserPatch,
+    db: db_dep,
+    user_token: user_admin_dep,
+    response: Response,
+):
     """Endpoint to partially update an existing user's information.
 
     Args:
         id (str): The unique identifier of the user to update.
         user (model.UserPatch): The updated user data.
         db (db_dep): The database session.
+        user_token (user_admin_dep): The user session with admin scope.
         response (Response): The FastAPI response object to set status codes.
 
     Returns:
         model.UserResponse: The updated user data.
     """
+    if user_token.scopes != "admin":
+        raise HTTPException(
+            status_code=status.HTTP_401_UNAUTHORIZED,
+            detail="Not enough permissions",
+            headers={"WWW-Authenticate": "Bearer scope=admin"},
+        )
     return service.patch_user(id, user, db, response)
 
 
 @router.delete("/{id}", status_code=204)
-def delete_user(id: str, db: db_dep, response: Response):
+def delete_user(
+    id: str, db: db_dep, user_token: user_admin_dep, response: Response
+):
     """Endpoint to delete a user by their user_id.
 
     Args:
         id (str): The unique identifier of the user to delete.
         db (db_dep): The database session.
+        user_token (user_admin_dep): The user session with admin scope.
         response (Response): The FastAPI response object to set status codes.
 
     Returns:
         None
     """
+    if user_token.scopes != "admin":
+        raise HTTPException(
+            status_code=status.HTTP_401_UNAUTHORIZED,
+            detail="Not enough permissions",
+            headers={"WWW-Authenticate": "Bearer scope=admin"},
+        )
     service.delete_user(id, db, response)
